ir-backend/requirements.txt
pandas~=2.2.2
fastapi~=0.111.0
pydantic~=2.7.4
numpy~=2.0.0
psycopg2~=2.9.9
python-dotenv~=1.0.1
